Move scripts to Pipeline/FeatureExtractor folder Add a description
