+++ Pipeline/FeatureExtractor/vggish_files/vggish_postprocess.py
@@ -17,7 +17,7 @@
 
 import numpy as np
 
-from vggish_files import vggish_params
+from Pipeline.FeatureExtractor.vggish_files import vggish_params
 
 
 class Postprocessor(object):
@@ -44,7 +44,7 @@ class Postprocessor(object):
     # Load means into a column vector for easier broadcasting later.
     self._pca_means = params[vggish_params.PCA_MEANS_NAME].reshape(-1, 1)
     assert self._pca_matrix.shape == (
-        vggish_params.EMBEDDING_SIZE, vggish_params.EMBEDDING_SIZE), (
+      vggish_params.EMBEDDING_SIZE, vggish_params.EMBEDDING_SIZE), (
             'Bad PCA matrix shape: %r' % (self._pca_matrix.shape,))
     assert self._pca_means.shape == (vggish_params.EMBEDDING_SIZE, 1), (
         'Bad PCA means shape: %r' % (self._pca_means.shape,))
@@ -82,9 +82,9 @@ class Postprocessor(object):
         vggish_params.QUANTIZE_MAX_VAL)
     # - convert to 8-bit in range [0.0, 255.0]
     quantized_embeddings = (
-        (clipped_embeddings - vggish_params.QUANTIZE_MIN_VAL) *
-        (255.0 /
-         (vggish_params.QUANTIZE_MAX_VAL - vggish_params.QUANTIZE_MIN_VAL)))
+            (clipped_embeddings - vggish_params.QUANTIZE_MIN_VAL) *
+            (255.0 /
+             (vggish_params.QUANTIZE_MAX_VAL - vggish_params.QUANTIZE_MIN_VAL)))
     # - cast 8-bit float to uint8
     quantized_embeddings = quantized_embeddings.astype(np.uint8)
